@@ -40,7 +40,7 @@ public class UndertaleSound : UndertaleNamedResource, INotifyPropertyChanged, ID
         /// <summary>
         /// Whether this sound uses the "new audio system".
         /// </summary>
-        /// <remarks>This is default for everything post Game Maker: Studio.
+        /// <remarks>This is default for everything post GameMaker Studio.
         /// The legacy sound system was used in pre Game Maker 8.</remarks>
         Regular = 0x64,
     }
@@ -85,7 +85,7 @@ public class UndertaleSound : UndertaleNamedResource, INotifyPropertyChanged, ID
     public UndertaleString File { get; set; }
 
     /// <summary>
-    /// A pre-Game Maker: Studio way of having certain effects on a sound effect.
+    /// A pre-GameMaker Studio way of having certain effects on a sound effect.
     /// </summary>
     /// <remarks>The exact way this works is unknown. But following values are possible:
     /// <c>Chorus</c>, <c>Echo</c>, <c>Flanger</c>, <c>Reverb</c>, <c>Gargle</c>, all possible to be combined with one another.</remarks>
@@ -144,6 +144,12 @@ public class UndertaleSound : UndertaleNamedResource, INotifyPropertyChanged, ID
     /// </summary>
     public int GroupID { get => _audioGroup.CachedId; set { _audioGroup.CachedId = value; OnPropertyChanged(); } }
 
+    /// <summary>
+    /// The precomputed length of the sound's audio data.
+    /// </summary>
+    /// <remarks>Introduced in GameMaker 2024.6.</remarks>
+    public float AudioLength { get; set; }
+
     /// <inheritdoc />
     public event PropertyChangedEventHandler PropertyChanged;
     
@@ -174,6 +180,9 @@ public class UndertaleSound : UndertaleNamedResource, INotifyPropertyChanged, ID
             writer.WriteUndertaleObject(_audioFile);
         else
             writer.Write(_audioFile.CachedId);
+
+        if (writer.undertaleData.IsVersionAtLeast(2024, 6))
+            writer.Write(AudioLength);
     }
 
     /// <inheritdoc />
@@ -207,6 +216,9 @@ public class UndertaleSound : UndertaleNamedResource, INotifyPropertyChanged, ID
         {
             _audioFile.CachedId = reader.ReadInt32();
         }
+
+        if (reader.undertaleData.IsVersionAtLeast(2024, 6))
+            AudioLength = reader.ReadSingle();
     }
 
     /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
