@@ -6,7 +6,7 @@ namespace UndertaleModLib.Models;
 /// A string entry a data file can have.
 /// </summary>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
-public class UndertaleString : UndertaleResource, ISearchable, IDisposable
+public class UndertaleString : UndertaleResource, ISearchable, IDisposable, Underanalyzer.IGMString
 {
     /// <summary>
     /// The contents of the string.
@@ -58,7 +58,7 @@ public class UndertaleString : UndertaleResource, ISearchable, IDisposable
     /// <summary>
     /// Returns a string that represents the current object.
     /// </summary>
-    /// <param name="isGMS2">Whether the string is from a Game Maker: Studio 2 data file or not.</param>
+    /// <param name="isGMS2">Whether the string is from a GameMaker Studio 2 data file or not.</param>
     /// <returns>A string that represents the current object.</returns>
     public string ToString(bool isGMS2)
     {
@@ -70,6 +70,8 @@ public class UndertaleString : UndertaleResource, ISearchable, IDisposable
             return "\"" + Content.Replace("\\", "\\\\").Replace("\r", "\\r").Replace("\n", "\\n").Replace("\"", "\\\"") + "\"";
 
         // Handle GM:S 1's lack of escaping
+        // Yes, in GM:S 1 you cannot escape quotation marks. You are expected to concatenate
+        // single-quote strings with double-quote strings.
         string res = Content;
         bool front, back;
         if (res.StartsWith('"'))
