@@ -1,4 +1,6 @@
-﻿using ICSharpCode.AvalonEdit;
+﻿#pragma warning disable CA1416 // Validate platform compatibility
+
+using ICSharpCode.AvalonEdit;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -66,5 +68,24 @@ namespace UndertaleModTool
 
             srcString.Content = editor.Text;
         }
+
+        private void TextEditor_DataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
+        {
+            var editor = sender as TextEditor;
+            if (editor is null)
+            {
+                return;
+            }
+
+            var srcString = e.NewValue as UndertaleString;
+            if (srcString is null)
+            {
+                return;
+            }
+
+            editor.Text = srcString.Content;
+        }
     }
 }
+
+#pragma warning restore CA1416 // Validate platform compatibility
