@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Diagnostics;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
@@ -53,22 +54,25 @@ namespace UndertaleModTool
             }
         }
 
+        public static bool ShowNullEntriesInResourceTree
+        {
+            get => Settings.Instance.ShowNullEntriesInResourceTree;
+            set
+            {
+                Settings.Instance.ShowNullEntriesInResourceTree = value;
+                Settings.Save();
+
+                // Refresh the tree for the change to take effect
+                mainWindow.UpdateTree();
+            }
+        }
+
         public static bool ProfileModeEnabled
         {
             get => Settings.Instance.ProfileModeEnabled;
             set
             {
                 Settings.Instance.ProfileModeEnabled = value;
-                Settings.Save();
-            }
-        }
-
-        public static bool UseGMLCache
-        {
-            get => Settings.Instance.UseGMLCache;
-            set
-            {
-                Settings.Instance.UseGMLCache = value;
                 Settings.Save();
             }
         }
@@ -99,6 +103,22 @@ namespace UndertaleModTool
             {
                 Settings.Instance.AutomaticFileAssociation = value;
                 Settings.Save();
+
+                if (!value && RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                {
+                    // Prompt user if they want to unassociate
+                    if (mainWindow.ShowQuestion("Remove current file associations, if they exist?", MessageBoxImage.Question, "File associations") == MessageBoxResult.Yes)
+                    {
+                        try
+                        {
+                            FileAssociations.RemoveAssociations();
+                        }
+                        catch (Exception ex)
+                        {
+                            mainWindow.ScriptError(ex.ToString(), "Unassociation failed", false);
+                        }
+                    }
+                }
             }
         }
 
@@ -181,6 +201,38 @@ namespace UndertaleModTool
             }
         }
 
+        public static string TransparencyGridColor1
+        {
+            get => Settings.Instance.TransparencyGridColor1;
+            set
+            {
+                try
+                {
+                    MainWindow.SetTransparencyGridColors(value, TransparencyGridColor2);
+
+                    Settings.Instance.TransparencyGridColor1 = value;
+                    Settings.Save();
+                }
+                catch (FormatException) { }
+            }
+        }
+
+        public static string TransparencyGridColor2
+        {
+            get => Settings.Instance.TransparencyGridColor2;
+            set
+            {
+                try
+                {
+                    MainWindow.SetTransparencyGridColors(TransparencyGridColor1, value);
+
+                    Settings.Instance.TransparencyGridColor2 = value;
+                    Settings.Save();
+                }
+                catch (FormatException) { }
+            }
+        }
+
         public static bool EnableDarkMode
         {
             get => Settings.Instance.EnableDarkMode;
@@ -209,11 +261,39 @@ namespace UndertaleModTool
             }
         }
 
+        public static bool RememberWindowPlacements
+        {
+            get => Settings.Instance.RememberWindowPlacements;
+            set
+            {
+                Settings.Instance.RememberWindowPlacements = value;
+                Settings.Save();
+            }
+        }
+
+        public static DecompilerSettings DecompilerSettings => Settings.Instance.DecompilerSettings;
+
+        public static string InstanceIdPrefix
+        {
+            get => Settings.Instance.InstanceIdPrefix;
+            set
+            {
+                Settings.Instance.InstanceIdPrefix = value;
+                Settings.Save();
+            }
+        }
+
         public bool UpdateButtonEnabled
         {
             get => UpdateAppButton.IsEnabled;
             set => UpdateAppButton.IsEnabled = value;
         }
+
+#if DEBUG
+        public static Visibility UpdaterButtonVisibility => Visibility.Visible;
+#else
+        public static Visibility UpdaterButtonVisibility => Visibility.Hidden;
+#endif
 
         public SettingsWindow()
         {
@@ -221,6 +301,7 @@ namespace UndertaleModTool
             this.DataContext = this;
             Settings.Load();
         }
+
         private void Window_IsVisibleChanged(object sender, DependencyPropertyChangedEventArgs e)
         {
             if (!IsVisible || IsLoaded)
@@ -238,6 +319,14 @@ namespace UndertaleModTool
         private void UpdateAppButton_Click(object sender, RoutedEventArgs e)
         {
             ((MainWindow)Owner).UpdateApp(this);
+        }
+
+        private void GMLSettingsButton_Click(object sender, RoutedEventArgs e)
+        {
+            GMLSettingsWindow settings = new(Settings.Instance);
+            settings.Owner = this;
+            settings.ShowDialog();
+            Settings.Save();
         }
     }
 }
